pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multidoc-summarizer"
version = "0.1.0"
description = "Agentic Multi-Document Summarization System"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    # Database
    "beanie>=1.25.0",
    "motor>=3.4.0",
    # AI / LLM / Agents
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.2.0",
    "openai>=1.25.0",
    # NLP / Summarization
    "transformers>=4.40.0",
    "torch>=2.2.0",
    "sentence-transformers>=3.0.0",
    "faiss-cpu>=1.8.0",
    # Document parsing
    "pypdf>=4.2.0",
    "python-docx>=1.1.0",
    "newspaper3k>=0.2.8",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.2.0",
    "httpx>=0.27.0",
    # Config
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["agents*", "conflict*", "summarizer*", "api*", "db*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
